Fix x1 and bwt.

+++ main/functions/TemporalFunctions.py
@@ -144,7 +144,7 @@ def x1(sub_signal, frequency):
     sub_signal = np.array(sub_signal)
     window_length = 2 * frequency
     aMs = 100
-    bP = np.ones(int(round(aMs * frequency) / 1000))
+    bP = np.ones(round(aMs * frequency / 1000))
 
     x_d = (sub_signal[1:] - sub_signal[0:len(sub_signal) - 1]) ** 2
     x_d = np.append(x_d, x_d[-1])
